Add AnyValue and replace ValueGenerator with AnyValue type
As part of the generalization of Values, we can remove the special cased
ValueGenerator and just replace it with a Value type. However, like
StringValue, we want to make it so you don't need to nest the values
unless you want to use a ref, so this introduces an AnyValue type, which
is a "typed, any Value".

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>